+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: ncofre <ncofre@student.42.us.org>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/07 23:40:19 by ncofre            #+#    #+#             */
-/*   Updated: 2020/11/07 23:48:17 by ncofre           ###   ########.fr       */
+/*   Updated: 2020/11/22 21:45:06 by ncofre           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < n)
